shopformer_2/requirements.txt
# Shopformer_2 Requirements
# For Apple Silicon (MPS) support

# Core ML
torch>=2.0.0
numpy>=1.24.0

# Configuration
pyyaml>=6.0

# Metrics and evaluation
scikit-learn>=1.3.0

# Visualization (optional, for evaluate.py --plot)
matplotlib>=3.7.0

# Logging
tensorboard>=2.14.0
